合法手の判定をした
